Added scheduled task to update positions nightly
