+++ config/routes.rb
@@ -10,8 +10,8 @@ Rails.application.routes.draw do
   root to: 'pages#home'
 
   require "sidekiq/web"
-  authenticate :user, lambda { |u| u.admin } do
+  # authenticate :user, lambda { |u| u.admin } do
     mount Sidekiq::Web => '/sidekiq'
-  end
+  # end
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end
